requirements.txt
fastapi
torch
torchvision
pillow